Renaming bo->entities and remove tests methods
