+++ src/main/java/com/ratepay/challenge/repository/BuyerRepository.java
@@ -1,6 +1,6 @@
 package com.ratepay.challenge.repository;
 
-import com.ratepay.challenge.bo.Buyer;
+import com.ratepay.challenge.entity.Buyer;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
