+++ src/main/java/com/ratepay/challenge/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package com.ratepay.challenge.repository;
 
-import com.ratepay.challenge.bo.Order;
+import com.ratepay.challenge.entity.Order;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
