+++ src/main/java/com/ratepay/challenge/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.ratepay.challenge.repository;
 
-import com.ratepay.challenge.bo.Product;
+import com.ratepay.challenge.entity.Product;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.UUID;
